re-organization continues to move functions out of av_spex_the_file
