+++ src/AV_Spex/utils/dir_setup.py
@@ -1,7 +1,9 @@
 import os
 import sys
 import shutil
+import re
 from ..utils.log_setup import logger
+from ..utils.find_config import config_path
 
 def validate_input_paths(input_paths, is_file_mode):
     source_directories = []
